WIP building auth and models for next gen pluto
