+++ common/src/main/scala/models/pluto/CommissionRecord.scala
@@ -1,8 +1,11 @@
 package models.pluto
 
+import java.sql.Timestamp
 import java.time.LocalDateTime
-import java.util.UUID
 
-case class CommissionRecord(collection_id:Option[Int], user:Option[Int], created:LocalDateTime, updated:LocalDateTime,
-                            gnm_commission_title:String, gnm_commission_status:String, gnm_commission_workinggroup:UUID,
-                            gnm_commission_description:Option[String], gnm_commission_owner:List[Int])
+case class CommissionRecord(id:Int, collectionId:Option[Int], siteId: Option[String], created: LocalDateTime, updated:LocalDateTime,
+                            title: String, status: String, description: Option[String], workingGroup: Int,
+                            originalCommissionerName:Option[String], scheduledCompletion:LocalDateTime, owner:String,
+                            notes:Option[String],
+                            productionOffice:String,
+                            originalTitle:Option[String])
